TTK-16756: Add list, insert and delete captions python scripts
